refactor: Move connect-nodejs login extension to plugin
