+++ index.js
@@ -77,6 +77,7 @@ cli.initialize = function (callback) {
     path.join(__dirname, 'lib', 'plugins', 'fs'),
     path.join(__dirname, 'lib', 'plugins', 'config'),
     path.join(__dirname, 'lib', 'plugins', 'issuers'),
+    path.join(__dirname, 'lib', 'plugins', 'login'),
     path.join(__dirname, 'lib', 'commands')
   ], callback)
 }
