+++ lib/commands/login.js
@@ -38,64 +38,6 @@ function registerLogin (cli, options, done) {
       var config = loadConfig(cli)
       var anvil = new AnvilConnect(config)
 
-      // add login method to anvil client
-      function login (email, password) {
-        var self = this
-
-        // construct the endpoint
-        // this one isn't included in openid-configuration
-        var uri = this.issuer + '/signin'
-
-        // authorization parameters
-        var params = this.authorizationParams({
-          provider: 'password',
-          email: email,
-          password: password
-        })
-
-        return new Promise(function (resolve, reject) {
-          request({
-            url: uri,
-            method: 'POST',
-            form: params,
-            headers: { 'referer': uri }
-          })
-          // If the server returns a 200 OK, then the server has responded that
-          // either:
-          //
-          //  - Bad username & password
-          //  - Other error related to sign-in
-          //
-          // Because the server does not return the error in a machine-friendly
-          // format (e.g. JSON) we just assume it is a bad username and password
-          //
-          // TODO: Consider implementing Resource Owner Password Credentials
-          .then(function (data) {
-            reject(new Error('Bad username or password'))
-          })
-          // this is so ugly, but redirects get treated as errors by HTTP clients.
-          // So we need to handle the expected result using `catch`. Ugh.
-          .catch(function (err) {
-            if (err.statusCode === 302) {
-              var u = url.parse(err.response.headers.location)
-              var code = qs.parse(u.query).code
-
-              self.token({ code: code })
-              .then(function (data) {
-                resolve(data)
-              })
-              .catch(function (err) {
-                reject(err)
-              })
-            } else {
-              reject(err)
-            }
-          })
-        })
-      }
-
-      AnvilConnect.prototype.login = login
-
       // prompt for email and password
       inquirer.prompt([
         {
